introducing controllers to clearn the routes file.
